some refactoring to simplify code
